part4 4.15-4.21

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogRouter = require('express').Router();
-const { response } = require('express');
+const jwt = require('jsonwebtoken');
+
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
 blogRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({});
@@ -19,14 +21,26 @@ blogRouter.get('/:id', async (req, res) => {
 blogRouter.post('/', async (req, res) => {
   const body = req.body;
 
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
   const blog = new Blog({
     title: body.title,
-    author: body.author,
+    author: body.author || user.name,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   res.json(savedBlog);
 });
 
@@ -48,7 +62,19 @@ blogRouter.put('/:id', async (req, res) => {
 });
 
 blogRouter.delete('/:id', async (req, res) => {
+  if (!req.token) {
+    return res.status(401).json({ error: 'token missing' });
+  }
+
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  const blog = await Blog.findById(req.params.id);
+
+  if (decodedToken.id !== blog.user.toString()) {
+    return res.status(401).json({ error: 'token invalid' });
+  }
+
   await Blog.findByIdAndRemove(req.params.id);
+
   res.status(204).end();
 });
